graph_lab/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-lab"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "surrealdb>=0.4",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
    "tiktoken>=0.7",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.40", "streamlit-agraph>=0.0.45"]
test = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
graph-lab = "graph_lab.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["graph_lab*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
